pyproject.toml
[project]
name = "chat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "fastapi",
    "uvicorn[standard]",
    "litellm",
    "python-dotenv",
    "aiosqlite"
]
